# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reco API",
        "description": "Reco - Offline-First Accounting SaaS Platform API Documentation",
        "contact": {
            "name": "Reco Support",
            "url": "https://reco.app/support",
            "email": "support@reco.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://reco.aaochaletaxi.app/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/dashboard/monthly-data": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get monthly income/expense data",
                "description": "Get monthly income and expense data for charts",
                "operationId": "getMonthlyData",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Year (default: current year)",
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monthly data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "months": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "Jan",
                                                        "Feb",
                                                        "Mar"
                                                    ]
                                                },
                                                "income": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "number"
                                                    },
                                                    "example": [
                                                        15000,
                                                        18000,
                                                        22000
                                                    ]
                                                },
                                                "expense": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "number"
                                                    },
                                                    "example": [
                                                        8000,
                                                        9500,
                                                        11000
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/receivables-trend": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get receivables trend",
                "description": "Get receivables trend data for the last N months",
                "operationId": "getReceivablesTrend",
                "parameters": [
                    {
                        "name": "months",
                        "in": "query",
                        "description": "Number of months (default: 6)",
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receivables trend data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "labels": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "Jan 2025",
                                                        "Feb 2025"
                                                    ]
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "number"
                                                    },
                                                    "example": [
                                                        45000,
                                                        52000
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/payables-trend": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get payables trend",
                "description": "Get payables trend data for the last N months",
                "operationId": "getPayablesTrend",
                "parameters": [
                    {
                        "name": "months",
                        "in": "query",
                        "description": "Number of months (default: 6)",
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payables trend data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "labels": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "Jan 2025",
                                                        "Feb 2025"
                                                    ]
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "number"
                                                    },
                                                    "example": [
                                                        25000,
                                                        28000
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/themes/toggle-dark-mode": {
            "post": {
                "tags": [
                    "Themes"
                ],
                "summary": "Toggle dark mode",
                "operationId": "toggleDarkMode",
                "responses": {
                    "200": {
                        "description": "Toggled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SuccessResponse": {
                "title": "SuccessResponse",
                "description": "Standard success response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "Standard error response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: 1|abc123...",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}